Test initial location
